requirements.txt
screen-brightness-control>=0.16.0
pystray>=0.19.0
Pillow>=9.0.0
keyboard>=0.13.5 